requirements.txt
openai==0.27.0
streamlit>=1.20.0
python-dotenv
pyairtable
